cast int to double
to workaround MSVC bug

+++ src/gm2_slha_io.hpp
@@ -117,7 +117,7 @@ Scalar GM2_slha_io::convert_to(const std::string& str)
    Scalar value;
    try {
       value = SLHAea::to<Scalar>(str);
-      if (!std::isfinite(value)) {
+      if (!std::isfinite(static_cast<double>(value))) {
          throw boost::bad_lexical_cast();
       }
    }  catch (const boost::bad_lexical_cast&) {
